+++ handlers/aws/handler.py
@@ -85,12 +85,14 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
         for replay_record in lambda_event["Records"]:
             event = json_parser(replay_record["body"])
             input_id = event["event_input_id"]
-            if input_id not in shipper_cache:
+            output_type = event["output_type"]
+            shipper_id = input_id + output_type
+            if shipper_id not in shipper_cache:
                 shipper = get_shipper_for_replay_event(
                     config=config,
-                    output_type=event["output_type"],
+                    output_type=output_type,
                     output_args=event["output_args"],
-                    event_input_id=event["event_input_id"],
+                    event_input_id=input_id,
                     replay_handler=replay_handler,
                 )
 
@@ -101,9 +103,9 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
                     )
                     continue
 
-                shipper_cache[input_id] = shipper
+                shipper_cache[shipper_id] = shipper
             else:
-                shipper = shipper_cache[input_id]
+                shipper = shipper_cache[shipper_id]
 
             shipper.send(event["event_payload"])
             replay_handler.add_event_id_with_receipt_handle(